A first cut at making `dcc` multiple files.
I propose the following straw-man split:

 * `compile`
 * the explanations
 * the gdb driver
 * `__main__`

Much of this could be decomposed further, or cut in different places.

I've opted to glue the pieces together as simply as possible: `cat`,
driven by a simple makefile.  More complexity would be nice, but Python
doesn't have anything like `#line`, which makes it harder to justify
(e.g.,) `cpp -traditional` or M4.

Fixes #6.
