+++ 200-explain.py
@@ -1,392 +1,3 @@
-#!/usr/bin/env python3
-
-# andrewt@unsw.edu.au May 2013-2017
-# compile a C program with runtime checking enabled
-# run-time errors are intercepted and an explanation
-# printed in a form comprehensible to beginner programmers
-# using	 information obtained from	gdb
-# clang errors are explained using matcher from https://github.com/cs50/help50
-# This program is executed in 4 ways
-#
-# 1) Invoked by the user from the command line to do the compilation
-#	 the binary produced has code added which intercepts runtime errors and runs this program
-#
-# 2) invoked from the C binary when a runtime error occurs - it then runs gdb
-#
-# 3) invoked via gdb to print details of the program state when the runtime error occurred
-#
-# 4) invoked from the C binary to watch for valgrind errors on stdin when using valgrind
-
-# The above 4 execution modes have been bundled into one file to simplify installation 
-
-# Code needs to be Python 2 compatible for old versions of GDB
-from __future__ import print_function
-
-import collections, os, platform, re, signal, subprocess, sys, traceback
-
-
-debug = int(os.environ.get('DCC_DEBUG', '0'))
-dcc_path = os.path.realpath(sys.argv[0])
-# replace CSE mount path with /home - otherwise breaks sometimes with ssh jobs
-dcc_path = re.sub(r'/tmp_amd/\w+/\w+ort/\w+/\d+/', '/home/', dcc_path)
-dcc_path = re.sub(r'^/tmp_amd/\w+/\w+ort/\d+/', '/home/', dcc_path)	
-dcc_path = re.sub(r'^/(import|export)/\w+/\d+/', '/home/', dcc_path)	
-
-colorize_output = sys.stderr.isatty() or os.environ.get('DCC_COLORIZE_OUTPUT', False)
-
-extra_c_compiler_args = "-std=gnu11 -g -lm -Wno-unused	-Wunused-comparison	 -Wunused-value -fno-omit-frame-pointer -fno-common -funwind-tables -fno-optimize-sibling-calls -Qunused-arguments".split()
-
-#
-# Compile the user's program adding some C code
-#
-def compile():
-	global colorize_output
-	print_explanations = True  
-	commandline_args = sys.argv[1:]
-	if not commandline_args:
-		print("Usage: %s [--valgrind|--memory|--no_explanation] <c-files>" % sys.argv[0], file=sys.stderr)
-		sys.exit(1)
-#	for c_compiler in ["clang-3.9", "clang-3.8", "clang"]:
-#		if search_path(c_compiler):	 # shutil.which not available in Python 2
-#			break
-	c_compiler = "clang"
-	which_sanitizer = "address"
-	incremental_compilation = False
-	leak_check = False
-	suppressions_file = os.devnull
-#	 linking_object_files = False
-	user_supplied_compiler_args = []
-	while commandline_args:
-		arg = commandline_args.pop(0)
-		if arg in ['-u', '--undefined', '--uninitialized', '--uninitialised', '-fsanitize=memory', '--memory']:
-			which_sanitizer = "memory"
-			continue
-		elif arg == '--valgrind':
-			which_sanitizer = "valgrind"
-			continue
-		elif arg == '--leak-check':
-			which_sanitizer = "valgrind"
-			leak_check = True
-			continue
-		elif arg.startswith('--suppressions='):
-			suppressions_file = arg[len('--suppressions='):]
-			continue
-		elif arg == '--explanation':
-			print_explanations = True
-			continue
-		elif arg == '--no_explanation':
-			print_explanations = False
-			continue
-		user_supplied_compiler_args.append(arg)
-		if arg	== '-c':
-			incremental_compilation = True
-#		 elif arg.endswith('.o'):
-#			 linking_object_files = True
-		elif arg == '-fcolor-diagnostics':
-			colorize_output = True
-		elif arg == '-fno-color-diagnostics':
-			colorize_output = False
-		elif arg == '-o' and commandline_args:
-			object_filename = commandline_args[0]
-			if object_filename.endswith('.c') and os.path.exists(object_filename):
-				print("%s: will not overwrite %s with machine code" % (os.path.basename(sys.argv[0]), object_filename), file=sys.stderr)
-				sys.exit(1)
-	source = dcc_c_source
-	source = source.replace('__DCC_PATH', '"'+dcc_path+'"')
-	source = source.replace('__DCC_SANITIZER', '"'+which_sanitizer+'"')
-	wrapper_source = dcc_wrapper_source
-	if which_sanitizer == "memory" and platform.architecture()[0][0:2] == '32':
-		if search_path('valgrind'):
-			# -fsanitize=memory requires 64-bits so we fallback to embedding valgrind
-			which_sanitizer = "valgrind"
-		else:
-			print("%s: uninitialized value checking not support on 32-bit architectures", file=sys.stderr)
-			sys.exit(1)
-	if which_sanitizer == "valgrind":
-		sanitizer_args = []
-		wrapper_source = valgrind_wrapper_source.replace('__DCC_MONITOR_VALGRIND', '"'+dcc_path+' --watch-stdin-for-valgrind-errors"')
-		wrapper_source = wrapper_source.replace('__DCC_LEAK_CHECK', "yes" if leak_check else "no")
-		wrapper_source = wrapper_source.replace('__DCC_SUPRESSIONS_FILE', suppressions_file)
-	elif which_sanitizer == "memory":
-		sanitizer_args = ['-fsanitize=memory']
-	else:
-		# fixme add code to check version supports these
-		sanitizer_args = ['-fsanitize=address', '-fsanitize=undefined', '-fno-sanitize-recover=undefined,integer']
-		which_sanitizer = "address"
-#	if c_compiler == 'clang-3.8' and platform.architecture()[0][0:2] == '32':
-#		# backport clang-3.8 on CSE	 machines needs this
-#		sanitizer_args +=  ['-target', 'i386-pc-linux-gnu']
-	# First run	 with -O enabled for better compile-time warnings 
-	command = [c_compiler, '-O', '-Wall']  + sanitizer_args + extra_c_compiler_args + user_supplied_compiler_args
-	if colorize_output:
-		command += ['-fcolor-diagnostics']
-	if debug:
-		print(" ".join(command), file=sys.stderr)
-	process = subprocess.run(command, stdin=subprocess.DEVNULL, stdout=subprocess.PIPE, stderr=subprocess.STDOUT, universal_newlines=True)
-	# workaround for  https://github.com/android-ndk/ndk/issues/184
-	if "undefined reference to `__" in process.stdout:
-			sanitizer_args = [c for c in sanitizer_args if not c in ['-fsanitize=undefined', '-fno-sanitize-recover=undefined,integer']]
-			command = [c_compiler, '-O', '-Wall']  + sanitizer_args + extra_c_compiler_args + user_supplied_compiler_args
-			if debug:
-				print("undefined reference to `__mulodi4'", file=sys.stderr)
-				print("recompiling", " ".join(command), file=sys.stderr)
-			process = subprocess.run(command, stdin=subprocess.DEVNULL, stdout=subprocess.PIPE, stderr=subprocess.STDOUT, universal_newlines=True)
-	if process.stdout:
-		output = process.stdout
-		# remove any ANSI codes
-		# http://stackoverflow.com/a/14693789
-		colourless_output = re.sub(r'\x1b[^m]*m', '', output)
-		lines = output.splitlines()
-		colourless_lines = colourless_output.splitlines()
-		errors_explained = 0
-		try:
-			i = 0
-			explanation_made = {}
-			last_explanation_file_line = ''
-			while i < len(lines) and print_explanations:
-				matching_error_messages = help_dcc(colourless_lines[i:]) or help(colourless_lines[i:])
-				if not matching_error_messages: 
-					break
-				matched_error_messages, explanation = matching_error_messages
-				# ignore some I don't know explanations
-				if re.match(r'not quite sure', explanation[0], flags=re.I):
-					break
-				explanation = "\n ".join([e for e in explanation if 'cs50' not in e])
-				explanation = explanation.replace("`clang`", 'the compiler')
-				# Don't repeat explanations
-				n_explained_lines = len(matched_error_messages)
-				# some help messages miss the caret
-				if (n_explained_lines == 1 or  n_explained_lines == 2) and len(lines) > i + 2 and has_caret(colourless_lines[i+2]):
-					n_explained_lines = 3
-				if len(lines) > i + n_explained_lines and re.match(r'^.*note:', colourless_lines[i + n_explained_lines]):
-					#print('note line detcted')
-					n_explained_lines += 1
-				e = re.sub('line.*', '', explanation, flags=re.I)
-				if e not in explanation_made:
-					explanation_made[e] = 1
-					m = re.match(r'^([^:]+:\d+):', colourless_lines[i])
-					if m:
-						if m.group(1) == last_explanation_file_line:
-							# stop if there are two errors for one line - the second is probably wrong
-							break
-						last_explanation_file_line = m.group(1)
-					print("\n".join(lines[i:i+n_explained_lines]),	file=sys.stderr)
-					if colorize_output:
-						print("\033[0m\033[34mEXPLANATION:\033[0m", explanation+'\n', file=sys.stderr)
-					else:
-						print("EXPLANATION:", explanation+'\n', file=sys.stderr)
-					if 'warning:' not in lines[i]:
-						errors_explained += 1
-				i += n_explained_lines
-				if errors_explained >= 3:
-					break
-			if not errors_explained:
-				sys.stderr.write('\n'.join(lines[i:])+"\n")
-		except Exception:
-			etype, evalue, etraceback = sys.exc_info()
-			eformatted = "\n".join(traceback.format_exception_only(etype, evalue))
-			print("%s: internal error: %s" % (os.path.basename(sys.argv[0]), eformatted), file=sys.stderr)
-			sys.stderr.write(output)
-		if process.returncode:
-			sys.exit(process.returncode)
-	# run a second time without -O for better debugging code
-	command = [c_compiler, '-w'] + sanitizer_args + extra_c_compiler_args + user_supplied_compiler_args
-	if incremental_compilation:
-		if debug:
-			print(" ".join(command), file=sys.stderr)
-		sys.exit(subprocess.call(command))
-	command +=	['-Wl,-wrap,main', '-x', 'c', '-']
-	if debug:
-		print(" ".join(command), file=sys.stderr)
-	if debug > 1:
-		print(" ".join(command), '<<eof', file=sys.stderr)
-		print(source + wrapper_source)
-		print("eof")
-	process = subprocess.run(command, input=source + wrapper_source, stdout=subprocess.PIPE, stderr=subprocess.STDOUT, universal_newlines=True)
-	# workaround for  https://github.com/android-ndk/ndk/issues/184
-	# can  be triggered	 not earlier
-	if "undefined reference to `__" in process.stdout:
-		command = [c for c in command if not c in ['-fsanitize=undefined', '-fno-sanitize-recover=undefined,integer']]
-		if debug:
-			print("undefined reference to `__mulodi4'", file=sys.stderr)
-			print("recompiling", " ".join(command), file=sys.stderr)
-		process = subprocess.run(command, input=source + wrapper_source, stdout=subprocess.PIPE, stderr=subprocess.STDOUT, universal_newlines=True)
-	print(process.stdout, end='', file=sys.stderr)
-	sys.exit(process.returncode)
-
-def search_path(program):
-	for path in os.environ["PATH"].split(os.pathsep):
-		full_path = os.path.join(path, program)
-		if os.path.isfile(full_path) and os.access(full_path, os.X_OK):
-			return full_path
-
-#
-# C code to intercept runtime errors and run this program
-#
-
-dcc_c_source = r"""
-#include <stdio.h>
-#include <unistd.h>
-#include <stdlib.h>
-#include <assert.h>
-#include <string.h>
-#include <signal.h>
-#include <limits.h>
-#include <errno.h>
-
-static int debug = 0;
-
-static void _dcc_exit(void) {
-	if (debug) fprintf(stderr, "_dcc_exit()\n");
-	// exit or _exit keeps executing sanitizer code - including perhaps superfluous output
-	// SIGPIPE avoids killed message from bash
-	signal(SIGPIPE, SIG_DFL);
-	kill(getpid(), SIGPIPE);
-	signal(SIGINT, SIG_DFL);
-	kill(getpid(), SIGINT);
-	// if SIGINT fails
-	kill(getpid(), SIGKILL);
-	// should never reach here
-	_exit(1);
-}
-
-static void setenvd(char *n, char *v) {
-	setenv(n, v, 1);
-	if (debug) fprintf(stderr, "setenv %s=%s\n", n, v);
-}
-
-static void _explain_error(void) {
-	// if a program has exhausted file descriptor then we need to close some to run gdb etc,
-	// so as a precaution we close a pile of file descriptiors which may or may not be open
-	for (int i = 4; i < 32; i++)
-		close(i);
-	if (debug) fprintf(stderr, "running %s\n", __DCC_PATH);
-	system(__DCC_PATH);
-	_dcc_exit();
-}
-
-static void _signal_handler(int signum) {
-	signal(SIGABRT, SIG_IGN);
-	signal(SIGSEGV, SIG_IGN);
-	signal(SIGINT, SIG_IGN);
-	signal(SIGXCPU, SIG_IGN);
-	signal(SIGXFSZ, SIG_IGN);
-	signal(SIGFPE, SIG_IGN);
-	signal(SIGILL, SIG_IGN);
-	char signum_buffer[1024];
-	sprintf(signum_buffer, "%d", (int)signum);
-	setenvd("DCC_SIGNAL", signum_buffer);
-	_explain_error();
-}
-
-
-void __dcc_start(void) __attribute__((constructor))
-#if __has_attribute(optnone)
- __attribute__((optnone))
-#endif
-;
-
-#define STACK_BYTES_TO_CLEAR 4096000
-void __dcc_start(void) {
-	// leave 0xbe for uninitialized variables which get fresh stack pages
-	char a[STACK_BYTES_TO_CLEAR];
-	memset(a, 0xbe, sizeof a);
-	debug = getenv("DCC_DEBUG") != NULL;
-	if (debug) fprintf(stderr, "__dcc_start\n");
-	debug = getenv("DCC_DEBUG") != NULL;
-	setenvd("DCC_SANITIZER", __DCC_SANITIZER);
-	setenvd("DCC_PATH", __DCC_PATH);
-
-	char pid_buffer[32];
-	snprintf(pid_buffer, sizeof pid_buffer, "%d", (int)getpid());
-	setenvd("DCC_PID", pid_buffer);
-	memset(pid_buffer, 0xbe, sizeof pid_buffer);
-	
-	signal(SIGABRT, _signal_handler);
-	signal(SIGSEGV, _signal_handler);
-	signal(SIGINT, _signal_handler);
-	signal(SIGXCPU, _signal_handler);
-	signal(SIGXFSZ, _signal_handler);
-	signal(SIGFPE, _signal_handler);
-	signal(SIGILL, _signal_handler);
-}
-
-// intercept ASAN explanation
-void _Unwind_Backtrace(void *a, ...) {
-	if (debug) fprintf(stderr, "_Unwind_Backtrace\n");
-	_explain_error();
-}
-
-
-// intercept ASAN explanation
-void __asan_on_error() {
-	if (debug) fprintf(stderr, "__asan_on_error\n");
-	setenvd("DCC_ASAN_ERROR", "1");
-	_explain_error();
-}
-
-char *__ubsan_default_options() {
-	return "verbosity=0:print_stacktrace=1:halt_on_error=1:detect_leaks=0";
-}
-
-char *__asan_default_options() {
-	return "verbosity=0:print_stacktrace=1:halt_on_error=1:detect_leaks=0:max_malloc_fill_size=4096000:quarantine_size_mb=16";
-}
-
-char *__msan_default_options() {
-	return "verbosity=0:print_stacktrace=1:halt_on_error=1:detect_leaks=0";
-}
-""" 
-
-dcc_wrapper_source = r""" 
-int __wrap_main(int argc, char *argv[], char *envp[]) {
-	extern int __real_main(int argc, char *argv[], char *envp[]);
-	int i;
-	char mypath[PATH_MAX];
-	realpath(argv[0], mypath);
-	setenvd("DCC_BINARY", mypath);
-	return __real_main(argc, argv, envp);
-}
-""" 
-
-valgrind_wrapper_source = r""" 
-int __wrap_main(int argc, char *argv[], char *envp[]) {
-	extern int __real_main(int argc, char *argv[], char *envp[]);
-	int i;
-	char mypath[PATH_MAX];
-	realpath(argv[0], mypath);
-	setenvd("DCC_BINARY", mypath);
-	int valgrind_running = getenv("DCC_VALGRIND_RUNNING") != NULL;
-	if (debug) printf("__wrap_main(valgrind_running=%d)\n", valgrind_running);
-	if (valgrind_running) {
-		// valgrind errors get reported earlier if we unbuffer stdout
-		// otherwise uninitialized variables may not be detected until fflush when program exits
-		// which produces poor error message
-		setbuf(stdout, NULL);		   
-		signal(SIGPIPE, SIG_DFL);
-		if (debug) printf("running __real_main\n");
-		return __real_main(argc, argv, envp);
-	}
-	if (debug) fprintf(stderr, "command=%s\n", __DCC_MONITOR_VALGRIND);
-	FILE *valgrind_error_pipe = popen(__DCC_MONITOR_VALGRIND, "w");
-	setbuf(valgrind_error_pipe, NULL);			
-	setenvd("DCC_VALGRIND_RUNNING", "1");
-	char fd_buffer[1024];
-	sprintf(fd_buffer, "--log-fd=%d", (int)fileno(valgrind_error_pipe));
-//	char *valgrind_command[] = {"/usr/bin/valgrind", "-q", "--vgdb=yes", "--max-stackframe=16000000", fd_buffer, "--track-origins=yes", "--vgdb-error=1"};
-	char *valgrind_command[] = {"/usr/bin/valgrind", "-q", "--vgdb=yes", "--leak-check=__DCC_LEAK_CHECK", "--suppressions=__DCC_SUPRESSIONS_FILE", "--max-stackframe=16000000", "--partial-loads-ok=no", fd_buffer, "--vgdb-error=1"};
-	int valgrind_command_len = sizeof valgrind_command / sizeof valgrind_command[0];
-	char *valgrind_argv[argc + 1 + valgrind_command_len];
-	for (i = 0; i < valgrind_command_len; i++)
-		valgrind_argv[i] = valgrind_command[i];
-	valgrind_argv[valgrind_command_len] = argv[0];
-	for (i = 1; i < argc; i++)
-		valgrind_argv[i+valgrind_command_len] = argv[i];
-	valgrind_argv[argc+valgrind_command_len] = NULL;
-	execvp("/usr/bin/valgrind", valgrind_argv);
-	return 0;
-}
-""" 
 #
 # some extra helpers 
 
@@ -1420,403 +1031,3 @@ def tilde_extract(lines):
 	return lines[0][start:start+length]
 
 # end help50 code
-
-#
-# Code below is executed from gdb.
-# It prints details of the program state likely to be of interest to
-# a beginner programmer
-#
-
-
-hash_define = collections.defaultdict(dict)
-source = {}
-
-def debug_print(level, *args, **kwargs):
-	if debug >= level:
-		kwargs['file'] = sys.stderr
-		print(*args, **kwargs)
-
-class Location():
-	def __init__(self, filename, line_number, column='', function='', params='', variable=''):
-		self.filename = filename
-		self.line_number = int(line_number)
-		self.column = column
-		self.function = function
-		self.params = params
-		self.variable = variable
-	def __str__(self):
-		return "Location(%s,%s,column=%s,function=%s,params=%s,variable=%s)" % (self.filename, self.line_number, self.column, self.function, self.params, self.variable)
-	def description(self):
-		params = self.params
-		if self.function == 'main' and params.startswith('argc=1,'):
-			params = ''
-		return "in %s(%s) at\n%s:%s: %s" % (self.function, params, self.filename, self.line_number, self.source_line())
-	def long_description(self):
-		params = self.params
-		if self.function == 'main' and params.startswith('argc=1,'):
-			params = ''
-		if colorize_output:
-			return "in %s(%s) in \033[31m%s\033[0m at \033[31mline %s\033[0m:\n\n" % (self.function, params, self.filename, self.line_number) + self.surrounding_source(markMiddle=True)
-		else:
-			return "in %s(%s) in %s at line %s:\n\n" % (self.function, params, self.filename, self.line_number) + self.surrounding_source(markMiddle=True)
-	def source_line(self, clean=False):
-		return fileline(self.filename, self.line_number, clean)
-	def surrounding_source(self, radius=2, clean=False, markMiddle=False):
-		source = ''
-		for offset in range(-radius, radius+1):
-			line = fileline(self.filename, self.line_number+offset, clean=clean)
-			if markMiddle and offset == 0:
-				line = re.sub(r'^ {0,3}', '-->', line)
-				if colorize_output:
-					line = '\033[31m' + line + '\033[0m'
-			source += line
-		return source
-	def is_user_location(self):
-		if not re.match(r'^[a-zA-Z]', self.function): return False 
-		if re.match(r'^/(usr|build)/', self.filename): return False 
-		if re.match(r'^\?', self.filename): return False 
-		return True
-	  
-	
-def fileline(filename, line_number, clean=False):
-	line_number = int(line_number)
-	if filename in source:
-		if line_number < 0 or line_number > len(source[filename]):
-			return ''
-		if clean:
-			return clean_c_source(source[filename][line_number - 1])
-		return source[filename][line_number - 1]
-	try:
-		with open(filename) as f:
-			source[filename] = f.readlines()
-			for line in source[filename]:
-				m = re.match(r'^\s*#\s*define\s*(\w+)\s*(.*\S)', line)
-				if m:
-					hash_define[filename][m.group(1)] = (line.rstrip(), m.group(2))
-		line = source[filename][line_number - 1].rstrip() + "\n"
-		if clean:
-			line = clean_c_source(line)
-		return line
-	except IOError:
-		debug_print(1, "fileline error can not open: %s" % (filename))
-	except IndexError:
-		pass
-	return ""
-
-# remove comments and truncate strings & character constants to zero-length
-def clean_c_source(c_source, leave_white_space=False):
-	c_source = re.sub("\\[\"']", '', c_source)
-	c_source = re.sub(r'".*?"', '', c_source)
-	c_source = re.sub(r"'.*?'", '', c_source)
-	c_source = re.sub(r'/[/\*].*', '', c_source)
-	if leave_white_space:
-		return c_source
-	return c_source.strip() + "\n"
-
-def gdb_execute(command):
-	debug_print(2, 'gdb.execute:', command)
-	str = gdb.execute(command, to_string=True)
-	debug_print(2, 'gdb.execute:', '->', str)
-	return str
-	
-def gdb_evaluate(expression):
-	debug_print(1, 'gdb_evaluate:', expression,)
-	value = gdb_execute('print %s' % expression)
-	value = re.sub('^[^=]*=\s*', '', value).strip()
-	debug_print(1, '->', value,)
-	return value.strip()
-
-def parse_gdb_stack_frame(line):
-	# note don't match function names starting with _ these are not user functions
-	m = re.match(
-		r'^\s*#(?P<frame_number>\d+)\s+(0x[0-9a-f]+\s+in+\s+)?'
-		r'(?P<function>[a-zA-Z][^\s\(]*).*\((?P<params>.*)\)\s+at\s+'
-		r'(?P<filename>[^\s:]+):(?P<line_number>\d+)\s*$',
-			line)
-	debug_print(1, 'parse_gdb_stack_frame', m != None, line)
-	if m:
-		filename = m.group('filename')
-		if filename.startswith("/usr/") or filename.startswith("../sysdeps/"): 
-			m = None
-	return m
-	
-def gdb_set_frame():
-	try:
-		stack = gdb_execute('where')
-		debug_print(1, "\nStack:\n",stack, "\n")
-		stack_lines = stack.splitlines()
-		frame = None
-		for line in stack_lines:
-			frame = parse_gdb_stack_frame(line)
-			if frame and os.path.exists(frame.group('filename')):
-				break
-		if not frame:
-			for line in reversed(stack_lines):
-				frame = parse_gdb_stack_frame(line) or frame
-		if not frame:
-			debug_print(1, 'gdb_set_frame no frame number')
-			return None
-		gdb_execute('frame %s' % frame.group('frame_number'))
-		return Location(frame.group('filename'), frame.group('line_number'), function=frame.group('function'), params=frame.group('params'))
-	except:
-		if debug: traceback.print_exc(file=sys.stderr)
-	
-def relevant_variables(c_source, arrays=[]):
-	expressions = extract_expressions(c_source)
-#	 arrays=[r'[a-z][a-zA-Z0-9_]*']
-#	 debug_print(1, 'relevant_variables', arrays, c_source)
-#	 for array in arrays:
-#		 indices = extract_indices(array, c_source)
-#		 expressions += indices
-	done = set(['NULL', 'char','int', 'double', 'while', 'if', 'else', 'for', 'while', 'return']) # avoid trying to evaluate types/keywords for efficency
-	explanation = ''
-	debug_print(1, 'relevant_variables expressions=', c_source, expressions)
-	for expression in sorted(expressions, key=lambda e: (len(re.findall(r'\w+', e)), e)):
-		try:
-			expression = expression.strip()
-			if expression not in done:
-				done.add(expression)
-				expression_value = evaluate_expression(expression)
-				if expression_value is not None:
-					explanation +=	"%s = %s\n" % (expression, expression_value)
-		except RuntimeError as e:
-			debug_print(1, 'print_variables_expressions: RuntimeError', e)
-			pass
-	if explanation:
-		prefix = 'Values when execution stopped:'
-		if colorize_output:
-			prefix = '\033[34m' + prefix + '\033[0m'
-		explanation = prefix + '\n\n' + explanation
-	return explanation
-
-def evaluate_expression(expression):
-	debug_print(1, 'expression:', expression)
-	if re.match(r'^-?\s*[\d\.]+$', expression):
-		return None	  # don't print(numbers)
-	if re.search(r'[a-zA-Z0-9_]\s*\(', expression):
-		return None	 # don't evaluate function calls
-		
-	expression_type = gdb_execute('whatis %s' % expression)
-	expression_type = re.sub(r'\s*type\s*=\s*', '',	 expression_type).strip()
-	debug_print(1, 'expresion_type=', expression_type)
-	if re.search(r'\<|\)$', expression_type):
-		return None
-
-	expression_value = gdb_evaluate(expression)
-
-	if (
-		'_IO_FILE' in expression_value or
-		'here_cg_arc_record' in	expression_value or
-		expression_value == '<optimized out>' or
-		len(expression_value) > 128
-		):
-		return None
-		
-	expression_value = re.sub(r'^0x[0-9a-f]+\s*(<.str>)?\s*"', '"', expression_value)
-	if re.search(r'^\(.*\s+0x[0-9a-f]{4,}\s*$', expression_value):
-		return None 
-	expression_value = re.sub(r'^\s*\(\S+\s+\*\)\s*0x0\s*$', 'NULL', expression_value)
-	if expression_type == 'char':
-		m = re.match(r"^(-?\d+) '(.*)'$", expression_value)
-		if m:
-			ascii = int(m.group(1))
-			if (0 < ascii < 9) or (13 < ascii < 32)	 or (ascii == 127):
-				 expression_value = '%d (non-printable ASCII character)' % ascii
-			elif ascii < 0 or ascii > 128:
-				 expression_value = '%d (not valid ASCII)' % ascii
-			elif ascii == 0:
-				 expression_value = "0 = '\\0'"
-			else:
-				 expression_value = "%s = '%s'" % m.groups()
-	expression_value = re.sub(r"'\000'", r"'\\0'", expression_value)
-	warning_text = " <-- warning appears to be uninitialized value"
-	if colorize_output:
-		warning_text = '\033[31m' + warning_text + '\033[0m'
-	for value in ['-1094795586', '-1.8325506472120096e-06', '-0.372548997', '-66 (not valid ASCII)']:
-		expression_value = expression_value.replace(value, value + warning_text)
-	return expression_value
-
-def balance_bracket(str, depth=0):
-#	 debug_print(1, 'balance_bracket(%s, %s)' % (str, depth))
-	if not str:
-		return ""
-	elif str[0] == ']' or str[0] == ')':
-		depth -= 1
-	elif str[0] == '[' or str[0] == '(':
-		depth += 1
-	if depth < 0 and (len(str) == 1 or str[1] != '['):
-		return ""
-	return str[0] +	 balance_bracket(str[1:], depth)
-
-	
-def extract_expressions(c_source):
-	expressions = []
-	# match declaration
-	m = re.match(r'[a-z][a-zA-Z0-9_]*\s+[a-z][a-zA-Z0-9_]*\s*\[(.*)', c_source, re.DOTALL)
-	if m:
-		 return extract_expressions(m.group(1))
-	m = re.match(r'([a-z][a-zA-Z0-9_]*)\s*\[(.*)', c_source, re.DOTALL)
-	if m:
-		index = balance_bracket(m.group(2))
-		if index:
-			expressions = [m.group(1), index, m.group(1) + '[' + index + ']']
-		remainder = m.group(2)
-	else:
-		m = re.match(r'([a-zA-Z][a-zA-Z0-9_]*)(.*)', c_source, re.DOTALL)
-		if m:
-			expressions = [m.group(1)]
-			remainder = m.group(2)
-		else:
-			m = re.match(r'^[^a-zA-Z]+(.*)', c_source, re.DOTALL)
-			if m:
-				remainder = m.group(1)
-			else:
-				return []
-	return expressions + extract_expressions(remainder)
-	
-def explain_signal(signal_number):
-	if signal_number == signal.SIGINT:
-		return "Execution was interrupted"
-	elif signal_number == signal.SIGFPE:
-		return 'Execution stopped by an arithmetic error.\nOften this is caused by division (or %) by zero.'
-	elif signal_number == signal.SIGXCPU:
-		return "Execution stopped by a CPU time limit."
-	elif signal_number == signal.SIGXFSZ:
-		return "Execution stopped because too much data written."
-	else:
-		return "Execution terminated by signal %s" % signal_number
-
-def explain_location(loc):
-	if not isinstance(loc, Location):
-		return "Execution stopped at '%s'" % (loc)
-	else:
-		return 'Execution stopped here ' + loc.long_description()
-
-def explain_error():
-	#subprocess.call("echo explain_error_gdb.py starting >/dev/tty", shell=True)
-	debug_print(1, 'explain_error() starting')
-	pid = None
-	# file descriptor 3 is a dup of stderr (see below)
-	# stdout & stderr have been diverted to /dev/null
-	output_stream = os.fdopen(3, "w")
-	try:
-		signal.signal(signal.SIGINT, handler)
-		pid = int(os.environ.get('DCC_PID', 0))
-		if 'DCC_VALGRIND_ERROR' in os.environ:
-			debug_print(1, 'attaching gdb to valgrind', pid)
-			gdb.execute('target remote | vgdb --pid=%d' % pid)
-		else:
-			debug_print(1, 'attaching gdb to ', pid)
-			gdb.execute('attach %s' % pid)
-		print(file=output_stream)
-		loc = gdb_set_frame()
-		signal_number = int(os.environ.get('DCC_SIGNAL', signal.SIGABRT))
-		if signal_number != signal.SIGABRT:
-			 print(explain_signal(signal_number), file=output_stream)
-		elif "DCC_ASAN_ERROR" in os.environ:
-			if loc:
-				print("%s:%d" % (loc.filename, loc.line_number), end=' ', file=output_stream)
-			print("runtime error - illegal array, pointer or other operation", file=output_stream)
-		elif os.environ.get('DCC_SANITIZER', '') == 'memory':
-			if loc:
-				print("%s:%d" % (loc.filename, loc.line_number), end=' ', file=output_stream)
-			print("runtime error - uninitialized variable used", file=output_stream)
-		if loc:
-			print(explain_location(loc), file=output_stream)
-			print(relevant_variables(loc.surrounding_source(clean=True)), file=output_stream)
-		gdb.flush(gdb.STDOUT)
-		gdb.flush(gdb.STDERR)
-	except gdb.error as e:
-		if 'ptrace' in str(e).lower() and os.path.exists('/.dockerenv'):
-			print('\ndcc : can not provide information about variables because docker not run with --cap-add=SYS_PTRACE\n' , file=output_stream)
-		elif debug:
-			traceback.print_exc(file=output_stream)
-		else:
-			print(e, file=output_stream)	
-		sys.exit(1)
-	except:
-		if debug: traceback.print_exc(file=output_stream)
-		sys.exit(1)
-
-#
-# ensure the program compiled with dcc terminates after error
-#
-def kill_program():
-	if 'DCC_PID' in os.environ:
-		try:
-			os.kill(int(os.environ['DCC_PID']), signal.SIGPIPE)
-			os.kill(int(os.environ['DCC_PID']), signal.SIGKILL)
-		except ProcessLookupError:
-			pass
-	sys.exit(1)
-
-def handler(signum, frame):
-	if debug: print >>sys.stderr, 'signal caught'
-	kill_program()
-	
-def run_gdb():
-	if colorize_output:
-		os.environ['DCC_COLORIZE_OUTPUT'] = 'true'
-	os.environ['DCC_RUN_INSIDE_GDB'] = 'true'
-	os.environ['PATH'] = '/bin:/usr/bin:/usr/local/bin:/sbin:/usr/sbin:' + os.environ.get('PATH', '')
-	os.environ['LC_ALL'] = 'C' # stop invalid utf-8 from gdb throwing Python exception
-	if not search_path('gdb'):
-		print('\ngdb not available to print variable values\n', file=sys.stderr)
-		kill_program()
-		return
-	command = ["gdb", "--nx", "--batch", "-ex", "python exec(open('%s').read())" % dcc_path, os.environ['DCC_BINARY']]
-	debug_print(1, 'running:', command)
-	# gdb puts confusing messages on stderr & stdout  so send these to /dev/null
-	# and use file descriptor 3 for our messages
-	os.dup2(2, 3)
-	if debug:
-		p = subprocess.Popen(command, stdin=subprocess.DEVNULL, close_fds=False)
-	else:
-		p = subprocess.Popen(command, stdin=subprocess.DEVNULL, stdout=subprocess.DEVNULL, stderr=subprocess.DEVNULL, close_fds=False)
-	p.communicate()
-	kill_program()
-if __name__ == '__main__':
-	signal.signal(signal.SIGINT, handler)
-	if debug: print(sys.argv, 'DCC_RUN_INSIDE_GDB="%s" DCC_SANITIZER="%s"' % (os.environ.get('DCC_RUN_INSIDE_GDB', ''), os.environ.get('DCC_PID', '')))
-	if not sys.argv[1:] and 'DCC_RUN_INSIDE_GDB' in os.environ:
-		# we are invoked by gdb 
-		import gdb
-		explain_error()
-		gdb_execute('call _exit()')
-		gdb_execute('quit')
-		kill_program()
-	elif not sys.argv[1:] and 'DCC_PID' in os.environ:
-		# we are invoked by the binary because an eror has occurred
-		run_gdb()
-	elif sys.argv[1:] == ['--watch-stdin-for-valgrind-errors']:
-		# valgrind is being used - we have been invoked viq the binary to watch for valgrind errors
-		# which have been directed to our stdin
-		while True:
-			line = sys.stdin.readline()
-			if not line:
-				break
-			debug_print(1, 'valgrind: ', line)
-			if 'vgdb me' in line:
-				if colorize_output:
-					os.environ['DCC_VALGRIND_ERROR'] = 'Runtime error: \033[31muninitialized variable accessed.\033[0m'
-				else:
-					os.environ['DCC_VALGRIND_ERROR'] = 'Runtime error: uninitialized variable accessed.'
-				print('\n'+os.environ['DCC_VALGRIND_ERROR'], file=sys.stderr)
-				sys.stderr.flush()
-				run_gdb()
-			elif 'loss record' in line:
-				line = sys.stdin.readline()
-				if 'malloc' in line:
-					line = sys.stdin.readline()
-					m = re.search(r'(\S+)\s*\((.+):(\d+)', line)
-					if m:
-						print('Error: free not called for memory allocated with malloc in function {} in {} at line {}.'.format(m.group(1),m.group(2),m.group(3)), file=sys.stderr)
-					else:
-						print('Error: free not called for memory allocated with malloc.', file=sys.stderr)
-				else:
-					print('Error: memory allocated not de-allocated.', file=sys.stderr)
-				sys.exit(0)
-	else:
-		# we are invoked by user to compile a program
-		os.environ['PATH'] = os.path.dirname(os.path.realpath(sys.argv[0])) + ':/bin:/usr/bin:/usr/local/bin:/sbin:/usr/sbin:' + os.environ.get('PATH', '') 
-		compile()
